requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
streamlit-extras>=0.3.0
google-cloud-documentai>=2.20.0
google-auth>=2.23.0
